Create pizza.py
In this exercise, I have to write a program from scratch. The program should prompt the user to input the information required:

The number of people.
The number of slices each person will eat.
The number of slices in each pizza pie.
Using that information, your program must calculate how many pizzas are needed to feed everyone.